+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.js
@@ -11,22 +11,13 @@ import { layout05, baseFontSize } from '@carbon/layout';
 import cx from 'classnames';
 import { useResizeDetector } from 'react-resize-detector';
 
-import {
-  BreadcrumbItem,
-  Grid,
-  Column,
-  Row,
-  Button,
-  Tag,
-} from 'carbon-components-react';
+import { Grid, Column, Row, Button, Tag } from 'carbon-components-react';
 
 import { useWindowResize, useNearestScroll } from '../../global/js/hooks';
 import { getDevtoolsProps } from '../../global/js/utils/devtools';
 
 import {
   deprecateProp,
-  deprecatePropUsage,
-  extractShapesArray,
   prepareProps,
 } from '../../global/js/utils/props-helper';
 
@@ -54,27 +45,19 @@ export let PageHeader = React.forwardRef(
       actionBarItems,
       actionBarMenuOptionsClass,
       actionBarOverflowAriaLabel,
-      actionBarOverflowLabel: deprecated_actionBarOverflowLabel,
       allTagsModalSearchLabel,
       allTagsModalSearchPlaceholderText,
       allTagsModalTitle,
-      availableSpace: deprecated_availableSpace,
-      background: deprecated_background,
-      hasBackgroundAlways,
+      hasBackgroundAlways: deprecated_hasBackgroundAlways,
       breadcrumbOverflowAriaLabel,
-      breadcrumbOverflowLabel: deprecated_breadcrumbOverflowLabel,
-      breadcrumbItems: deprecated_breadcrumbItems,
-      breadcrumbs: breadcrumbsIn,
+      breadcrumbs,
       children,
       className,
       collapseHeader,
       collapseHeaderIconDescription,
-      collapseHeaderLabel: deprecated_collapseHeaderLabel,
-      collapseHeaderToggleWanted: deprecated_collapseHeaderToggleWanted,
       collapseTitle,
       disableBreadcrumbScroll,
       expandHeaderIconDescription,
-      expandHeaderLabel: deprecated_expandHeaderLabel,
       fullWidthGrid,
       hasCollapseHeaderToggle,
       narrowGrid,
@@ -82,28 +65,18 @@ export let PageHeader = React.forwardRef(
       pageActions,
       pageActionsOverflowLabel,
       pageActionsMenuOptionsClass,
-      pageHeaderOffset: _deprecated_pageHeaderOffset,
-      preCollapseTitleRow: deprecated_preCollapseTitleRow,
-      preventBreadcrumbScroll: deprecated_preventBreadcrumbScroll,
       showAllTagsLabel,
       subtitle,
       tags,
       title,
+      withoutBackground,
       ...rest
     },
     ref
   ) => {
     // handle deprecated props - START
-    actionBarOverflowAriaLabel ??= deprecated_actionBarOverflowLabel;
-    breadcrumbOverflowAriaLabel ??= deprecated_breadcrumbOverflowLabel;
-    children ??= deprecated_availableSpace;
-    collapseHeaderIconDescription ??= deprecated_collapseHeaderLabel;
-    expandHeaderIconDescription ??= deprecated_expandHeaderLabel;
-    hasBackgroundAlways ??= deprecated_background;
-    hasCollapseHeaderToggle ??= deprecated_collapseHeaderToggleWanted;
-    collapseTitle ??= deprecated_preCollapseTitleRow;
-    disableBreadcrumbScroll ??= deprecated_preventBreadcrumbScroll;
-    const breadcrumbs = breadcrumbsIn ?? deprecated_breadcrumbItems;
+    // if withoutBackground is null check deprecated_hasBackgroundAlways and default false
+    withoutBackground ??= !(deprecated_hasBackgroundAlways ?? true);
     // handle deprecated props - END
 
     const [metrics, setMetrics] = useState({});
@@ -129,8 +102,7 @@ export let PageHeader = React.forwardRef(
     };
 
     // state based on props only
-    const actionBarItemArray = extractShapesArray(actionBarItems);
-    const hasActionBar = actionBarItemArray && actionBarItemArray.length > 0;
+    const hasActionBar = actionBarItems && actionBarItems.length > 0;
     const hasBreadcrumbRow = breadcrumbs || actionBarItems;
 
     // NOTE: The buffer is used to add space between the bottom of the header and the last content
@@ -375,8 +347,8 @@ export let PageHeader = React.forwardRef(
     }, [fullWidthGrid, narrowGrid]);
 
     useEffect(() => {
-      // Determines if the hasBackgroundAlways should be one based on the header height or scroll
-      let result = hasBackgroundAlways ? 1 : 0;
+      // Determines the appropriate header background opacity based on the header config/height/scroll and the withoutBackground setting
+      let result = withoutBackground ? 0 : 1;
 
       if (
         !result &&
@@ -403,7 +375,7 @@ export let PageHeader = React.forwardRef(
       }));
     }, [
       actionBarItems,
-      hasBackgroundAlways,
+      withoutBackground,
       breadcrumbs,
       headerRef,
       metrics.breadcrumbRowHeight,
@@ -415,11 +387,11 @@ export let PageHeader = React.forwardRef(
     ]);
 
     useEffect(() => {
-      // only has toggle if requested and has hasBackgroundAlways
+      // only has toggle if requested and withoutBackground is unset/falsy
       // NOTE: prop-types isRequired.if for the expand and collapse
       // icon descriptions depends on the this.
-      setHasCollapseButton(hasCollapseHeaderToggle && hasBackgroundAlways);
-    }, [hasBackgroundAlways, hasCollapseHeaderToggle]);
+      setHasCollapseButton(hasCollapseHeaderToggle && !withoutBackground);
+    }, [withoutBackground, hasCollapseHeaderToggle]);
 
     useEffect(() => {
       // Determine if space is needed in the breadcrumb for a collapse button
@@ -523,22 +495,13 @@ export let PageHeader = React.forwardRef(
                           noTrailingSlash={!!title}
                           overflowAriaLabel={breadcrumbOverflowAriaLabel}
                           breadcrumbs={
-                            breadcrumbsIn
+                            breadcrumbs
                               ? breadcrumbItemForTitle
-                                ? breadcrumbsIn.concat(breadcrumbItemForTitle)
-                                : breadcrumbsIn
+                                ? breadcrumbs.concat(breadcrumbItemForTitle)
+                                : breadcrumbs
                               : null
                           }
-                        >
-                          {!breadcrumbsIn ? deprecated_breadcrumbItems : null}
-                          {!breadcrumbsIn && breadcrumbItemForTitle ? (
-                            <BreadcrumbItem
-                              {...prepareProps(breadcrumbItemForTitle, 'label')}
-                            >
-                              {breadcrumbItemForTitle.label}
-                            </BreadcrumbItem>
-                          ) : null}
-                        </BreadcrumbWithOverflow>
+                        />
                       ) : null}
                     </Column>
                     <Column
@@ -566,7 +529,7 @@ export let PageHeader = React.forwardRef(
                                 `${blockClass}__action-bar-menu-options`
                               )}
                               overflowAriaLabel={actionBarOverflowAriaLabel}
-                              actions={actionBarItemArray}
+                              actions={actionBarItems}
                               className={`${blockClass}__action-bar`}
                               onWidthChange={handleActionBarWidthChange}
                               rightAlign={true}
@@ -794,81 +757,11 @@ const tagTypes = Object.keys(TYPES);
 
 export const deprecatedProps = {
   /**
-   * **Deprecated** see property `actionBarOverflowAriaLabel`
+   * **Deprecated** see property `withoutBackground`
    */
-  actionBarOverflowLabel: deprecateProp(
-    PropTypes.string,
-    'Property renamed to `actionBarOverflowAriaLabel`.'
-  ),
-  /**
-   * **Deprecated** see property `children`
-   */
-  availableSpace: deprecateProp(
-    PropTypes.node,
-    'Make use of children instead.'
-  ),
-  /**
-   * **Deprecated** see property `hasBackgroundAlways`
-   */
-  background: deprecateProp(
+  hasBackgroundAlways: deprecateProp(
     PropTypes.bool,
-    'Property renamed to `hasBackgroundAlways`'
-  ),
-  /**
-   * **Deprecated** see property `breadcrumbs`
-   */
-  breadcrumbItems: deprecateProp(
-    PropTypes.element,
-    'Usage changed to expect breadcrumb item like shapes, see `breadcrumbs`.'
-  ),
-  /**
-   * **Deprecated** see property `breadcrumbOverflowAriaLabel`
-   */
-  breadcrumbOverflowLabel: deprecateProp(
-    PropTypes.string,
-    'Property renamed to `breadcrumbOverflowAriaLabel`.'
-  ),
-  /**
-   * **Deprecated** see property `collapseHeaderIconDescription`
-   */
-  collapseHeaderLabel: deprecateProp(
-    PropTypes.string,
-    'Property renamed to `collapseHeaderIconDescription`.'
-  ),
-  /**
-   * **Deprecated** see property `hasCollapseHeaderToggle`
-   */
-  collapseHeaderToggleWanted: deprecateProp(
-    PropTypes.bool,
-    'Property renamed to `hasCollapseHeaderToggle`'
-  ),
-  /**
-   * **Deprecated** see property `expandHeaderIconDescription`
-   */
-  expandHeaderLabel: deprecateProp(
-    PropTypes.string,
-    'Property renamed to `expandHeaderIconDescription`.'
-  ),
-  /**
-   * **Deprecated** no longer required
-   */
-  pageHeaderOffset: deprecateProp(
-    PropTypes.number,
-    'Property removed as no longer required.'
-  ),
-  /**
-   * **Deprecated** see property `collapseTitle`
-   */
-  preCollapseTitleRow: deprecateProp(
-    PropTypes.bool,
-    'Property renamed to `collapseTitle`.'
-  ),
-  /**
-   * **Deprecated** see property `disableBreadcrumbScroll`
-   */
-  preventBreadcrumbScroll: deprecateProp(
-    PropTypes.bool,
-    'Prop renamed to `disableBreadcrumbScroll`.'
+    'Property replaced by `withoutBackground`'
   ),
 };
 
@@ -878,27 +771,18 @@ PageHeader.propTypes = {
    * Each item is specified as an object with the properties of a Carbon Button in icon only form.
    * Button kind, size, tooltipPosition, tooltipAlignment and type are ignored.
    */
-  actionBarItems: deprecatePropUsage(
-    PropTypes.arrayOf(
-      PropTypes.shape({
-        ...prepareProps(Button.propTypes, [
-          'kind',
-          'size',
-          'tooltipPosition',
-          'tooltipAlignment',
-        ]),
-        iconDescription: PropTypes.string.isRequired,
-        onClick: Button.propTypes.onClick,
-        renderIcon: Button.propTypes.renderIcon.isRequired,
-      })
-    ),
-
-    // expects action bar item as array or in fragment
-    PropTypes.oneOfType([
-      PropTypes.arrayOf(PropTypes.element),
-      PropTypes.element,
-    ]),
-    'Expects an array of objects with the following properties: iconDescription, renderIcon and onClick.'
+  actionBarItems: PropTypes.arrayOf(
+    PropTypes.shape({
+      ...prepareProps(Button.propTypes, [
+        'kind',
+        'size',
+        'tooltipPosition',
+        'tooltipAlignment',
+      ]),
+      iconDescription: PropTypes.string.isRequired,
+      onClick: Button.propTypes.onClick,
+      renderIcon: Button.propTypes.renderIcon.isRequired,
+    })
   ),
   /**
    * class name applied to the action bar overflow options
@@ -911,8 +795,7 @@ PageHeader.propTypes = {
    * NOTE: This prop is required if actionBarItems are supplied
    */
   actionBarOverflowAriaLabel: PropTypes.string.isRequired.if(
-    ({ actionBarItems, actionBarOverflowLabel }) =>
-      actionBarItems && actionBarItems.length > 0 && !actionBarOverflowLabel
+    ({ actionBarItems }) => actionBarItems && actionBarItems.length > 0
   ),
   /**
    * When tags are supplied there may not be sufficient space to display all of the tags. This results in an overflow
@@ -943,9 +826,7 @@ PageHeader.propTypes = {
    * It is used in an overflow menu when there is insufficient space to display all breadcrumbs inline.
    */
   breadcrumbOverflowAriaLabel: PropTypes.string.isRequired.if(
-    ({ breadcrumbs, breadcrumbItems }) =>
-      (breadcrumbs && breadcrumbs.length > 0) ||
-      (breadcrumbItems && breadcrumbItems.length > 0)
+    ({ breadcrumbs }) => breadcrumbs && breadcrumbs.length > 0
   ),
   /**
    * Specifies the breadcrumb components to be shown in the breadcrumb area of
@@ -1007,12 +888,12 @@ PageHeader.propTypes = {
    */
   collapseHeader: PropTypes.bool,
   /**
-   * If `hasCollapseHeaderToggle` and `hasBackgroundAlways` are set then assistive text is required
-   * for both the expend and collapse states of the button component used.
+   * If `hasCollapseHeaderToggle` is set and `withoutBackground` is unset/falsy then assistive text is
+   * required for both the expend and collapse states of the button component used.
    */
   collapseHeaderIconDescription: PropTypes.string.isRequired.if(
-    ({ hasBackgroundAlways, hasCollapseHeaderToggle }) =>
-      hasBackgroundAlways && hasCollapseHeaderToggle
+    ({ withoutBackground, hasCollapseHeaderToggle }) =>
+      !withoutBackground && hasCollapseHeaderToggle
   ),
   /**
    * The title row typically starts below the breadcrumb row. This option
@@ -1025,23 +906,18 @@ PageHeader.propTypes = {
    */
   disableBreadcrumbScroll: PropTypes.bool,
   /**
-   * If `hasCollapseHeaderToggle` and `hasBackgroundAlways` are set then assistive text is required
-   * for both the expend and collapse states of the button component used.
+   * If `hasCollapseHeaderToggle` is set and `withoutBackground` is unset/falsy then assistive text is
+   * required for both the expend and collapse states of the button component used.
    */
   expandHeaderIconDescription: PropTypes.string.isRequired.if(
-    ({ hasBackgroundAlways, hasCollapseHeaderToggle }) =>
-      hasBackgroundAlways && hasCollapseHeaderToggle
+    ({ withoutBackground, hasCollapseHeaderToggle }) =>
+      !withoutBackground && hasCollapseHeaderToggle
   ),
   /**
    * The PageHeader is hosted in a Carbon grid, this value is passed through to the Carbon grid fullWidth prop.
    * 'xl' is used to override the grid width setting. Can be used with narrowGrid: true to get the largest size.
    */
   fullWidthGrid: PropTypes.oneOfType([PropTypes.bool, PropTypes.oneOf(['xl'])]),
-  /**
-   * Specifies if the PageHeader should have a background always on and defaults to the preferred `true`.
-   * When false some parts of the header gain a background if they stick to the top of the PageHeader on scroll.
-   */
-  hasBackgroundAlways: PropTypes.bool,
   /**
    * Adds a button as the last element of the bottom row which collapses and expands the header.
    *
@@ -1169,12 +1045,16 @@ PageHeader.propTypes = {
       asText: PropTypes.string.isRequired,
     }),
   ]),
+  /**
+   * Specifies if the PageHeader should appear without a background color, and defaults to the preferred `false` (a background color is shown).
+   * Note that when `true` some parts of the header still gain a background if and when they stick to the top of the PageHeader on scroll.
+   */
+  withoutBackground: PropTypes.bool,
   ...deprecatedProps,
 };
 
 PageHeader.defaultProps = {
   fullWidthGrid: false,
-  hasBackgroundAlways: true,
   narrowGrid: false,
 };